requirements.txt
alchemical==0.7.1
alembic==1.12.0
apifairy==1.3.0
apispec==6.3.0
blinker==1.6.2
certifi==2023.7.22
click==8.1.7
elastic-transport==8.4.1
elasticsearch==8.10.0
exceptiongroup==1.1.3
Faker==19.6.2
flake8==6.1.0
flask==3.0.0
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
flask-sqlalchemy==3.1.1
greenlet==3.0.0
gunicorn==21.2.0
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mccabe==0.7.0
packaging==23.2
pluggy==1.3.0
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.21
tomli==2.0.1
typing-extensions==4.8.0
urllib3==1.26.17
watchdog==3.0.0
webargs==8.3.0
werkzeug==3.0.0
zipp==3.17.0
